pr small fixes
