--- webapp/packages/plugin-script-export/src/ExportScriptDialog/ExportScriptDialog.module.css
@@ -1,17 +1,0 @@
-/*
- * CloudBeaver - Cloud Database Manager
- * Copyright (C) 2020-2025 DBeaver Corp and others
- *
- * Licensed under the Apache License, Version 2.0.
- * you may not use this file except in compliance with the License.
- */
-
-.tabList {
-  composes: theme-border-color-background from global;
-  flex-shrink: 0;
-  padding: 0 12px;
-}
-
-.footerSlot {
-  display: contents;
-}
