--- webapp/packages/plugin-script-export/src/IScriptExportTabProps.ts
@@ -1,15 +1,0 @@
-/*
- * CloudBeaver - Cloud Database Manager
- * Copyright (C) 2020-2026 DBeaver Corp and others
- *
- * Licensed under the Apache License, Version 2.0.
- * you may not use this file except in compliance with the License.
- */
-
-export interface IScriptExportTabProps {
-  script: string;
-  fileName: string;
-  editorId: string;
-  projectId?: string;
-  connectionId?: string;
-}
