+++ webapp/packages/plugin-script-export/src/locales/it.ts
@@ -1,9 +1,1 @@
-/*
- * CloudBeaver - Cloud Database Manager
- * Copyright (C) 2020-2025 DBeaver Corp and others
- *
- * Licensed under the Apache License, Version 2.0.
- * you may not use this file except in compliance with the License.
- */
-
 export default [['plugin_script_export_dialog_title', 'Esporta script SQL']];
