+++ webapp/packages/plugin-sql-editor/src/LocalExport/LocalExportTab.tsx
@@ -5,20 +5,15 @@
  * Licensed under the Apache License, Version 2.0.
  * you may not use this file except in compliance with the License.
  */
-import { observable } from 'mobx';
 import { observer } from 'mobx-react-lite';
 
-import { Button, Fill, Form, InputField, Translate, useFocus, useForm, useObservableRef } from '@cloudbeaver/core-blocks';
+import { Button, Fill, Form, InputField, Translate, useFocus, useForm } from '@cloudbeaver/core-blocks';
 import { useScriptExportDialog, type IScriptExportTabProps } from '@cloudbeaver/plugin-script-export';
 import type { TabContainerPanelComponent } from '@cloudbeaver/core-ui';
 import { downloadSql } from '../downloadSql.js';
 import { NotificationService } from '@cloudbeaver/core-events';
 import { useService } from '@cloudbeaver/core-di';
-
-interface State {
-  fileName: string;
-  export: () => string;
-}
+import { useState } from 'react';
 
 export const LocalExportTab: TabContainerPanelComponent<IScriptExportTabProps> = observer(function LocalExportTab({
   script,
@@ -27,33 +22,25 @@ export const LocalExportTab: TabContainerPanelComponent<IScriptExportTabProps> =
   const { resolveDialog, rejectDialog, FooterSlot } = useScriptExportDialog();
   const [focusedRef] = useFocus<HTMLFormElement>({ focusFirstChild: true });
   const notificationService = useService(NotificationService);
+  const [fileName, setFileName] = useState(initialFileName);
 
-  const state = useObservableRef<State>(
-    () => ({
-      fileName: initialFileName,
-    }),
-    {
-      fileName: observable.ref,
-    },
-    false,
-  );
   const form = useForm({
     onSubmit() {
-      downloadSql(state.fileName, script);
+      downloadSql(fileName, script);
       resolveDialog();
       notificationService.logSuccess({
         title: 'sql_editor_script_exported',
-        message: state.fileName,
+        message: fileName,
       });
     },
   });
 
-  const canSave = state.fileName.trim().length > 0;
+  const canSave = fileName.trim().length > 0;
 
   return (
     <>
       <Form ref={focusedRef} context={form} className="tw:p-6 tw:w-full">
-        <InputField name="fileName" state={state} required small>
+        <InputField name="fileName" value={fileName} onChange={setFileName} required small>
           <Translate token="ui_file_name" />
         </InputField>
       </Form>
